clienterrors: make `Error.Details` a []error
Instead of allowing any value value to be passing into `Details`
use a `[]error`. This allows us to get compile time checking that
we pass "expected" values in. By being too lose we risk that we
pass values that cannot be json.Serialized. We could make the
type also `json.Marshaler` but the downside here is that we
would have to implement a bunch of custom marshalers. While
more generic it's also more work (that seems to be not needed
as we deal with errors anyway).