vendor: pull in github.com/osbuild/images/data/repositories
This commit just runs `go vendor` and `git add`s the new content.
It is required because in the previous commit we make use of the
"real" repositories in the `compose_test.go` which mean they need
to get added now (we only pulled in the test repos so far).